fix:修复优雅停机问题  - 修复注册中心 unexport 的并发问题  - 修复关闭过程中，拿到的线程池为 null 的问题  - 修复关闭过程中，拿到的 futures 为 null 的问题

+++ dubbo-config/dubbo-config-api/src/main/java/com/alibaba/dubbo/config/ProtocolConfig.java
@@ -29,6 +29,7 @@ import com.alibaba.dubbo.remoting.telnet.TelnetHandler;
 import com.alibaba.dubbo.rpc.Protocol;
 
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * ProtocolConfig
@@ -124,6 +125,8 @@ public class ProtocolConfig extends AbstractConfig {
     // 是否为缺省
     private Boolean isDefault;
 
+    private static final AtomicBoolean destroyed = new AtomicBoolean(false);
+
     public ProtocolConfig() {
     }
 
@@ -136,7 +139,11 @@ public class ProtocolConfig extends AbstractConfig {
         setPort(port);
     }
 
+    // TODO: 2017/8/30 to move this method somewhere else
     public static void destroyAll() {
+        if (!destroyed.compareAndSet(false, true)) {
+            return;
+        }
         AbstractRegistryFactory.destroyAll();
         ExtensionLoader<Protocol> loader = ExtensionLoader.getExtensionLoader(Protocol.class);
         for (String protocolName : loader.getLoadedExtensions()) {
@@ -439,7 +446,6 @@ public class ProtocolConfig extends AbstractConfig {
     public void destory() {
         if (name != null) {
             ExtensionLoader.getExtensionLoader(Protocol.class).getExtension(name).destroy();
-            ;
         }
     }
